Rename module, add initial tutorial file
