+++ doc/reference/handlers.rst
@@ -12,4 +12,4 @@ Functions for handling bulk data elements such as (7FE0,0010) *Pixel Data* and
 
    handlers.overlay_data
    handlers.pixel_data
-   handlers.waveform_data
+   handlers.waveforms
