+++ doc/tutorials/index.rst
@@ -14,6 +14,6 @@ New to *pydicom*? Then these tutorials should get you up and running.
    dataset_basics
    sr_basics
    dicom_json
-   waveform
+   waveforms
    contributing_code
    contributing_docs
