+++ pydicom/waveforms/__init__.py
@@ -1,0 +1,3 @@
+from pydicom.waveforms.numpy_handler import (
+    generate_multiplex, multiplex_array
+)
