+++ pydicom/waveforms/numpy_handler.py
